Data through October 19 (see note)
In order to get the deltas on the NA days, the script averages the surrounding PBA-level data.  This can be zeroed manually by inserting NAs in both the day and the day after there was no data in the amba_pba_sep.csv.

The script then subtracts the conurbano figure from PBA total to get the "PBA minus conurbano" number and adds the conurbano number to CABA.  Since the conurbano number is estimated for days in which there are gaps, there may be inaccuracies on and around those days.